QR code generator
This program creates a QR code for a text or URL of your choice and is saved in the same folder as the code is kept. According to Oxford Languages, a QR code is "a machine-readable code consisting of an array of black and white squares, typically used for storing URLs or other information for reading by the camera on a smartphone".